+++ noodles-tabix/src/lib.rs
@@ -29,32 +29,6 @@ pub mod fs;
 pub mod index;
 pub mod io;
 
-#[deprecated(since = "0.48.0", note = "Use `tabix::fs::read` instead.")]
-pub use self::fs::read;
-
-#[deprecated(since = "0.48.0", note = "Use `tabix::fs::write` instead.")]
-pub use self::fs::write;
-
-#[deprecated(since = "0.45.0", note = "Use `noodles_tabix::io::Reader` instead.")]
-pub use self::io::Reader;
-
-#[deprecated(since = "0.45.0", note = "Use `noodles_tabix::io::Writer` instead.")]
-pub use self::io::Writer;
-
-#[cfg(feature = "async")]
-#[deprecated(
-    since = "0.45.0",
-    note = "Use `noodles_tabix::r#async::io::Reader` instead."
-)]
-pub use self::r#async::Reader as AsyncReader;
-
-#[cfg(feature = "async")]
-#[deprecated(
-    since = "0.45.0",
-    note = "Use `noodles_tabix::r#async::io::Writer` instead."
-)]
-pub use self::r#async::Writer as AsyncWriter;
-
 use noodles_csi::binning_index::{self, index::reference_sequence::index::LinearIndex};
 
 const MAGIC_NUMBER: [u8; 4] = *b"TBI\x01";